+++ blueMSX/Src/Board/Machine.c
@@ -206,7 +206,7 @@ char *strcasestr(const char *str1, const char *str2)
 		return NULL; // No match
 
 	// Compute the offset in the lowercased version
-	offset = ptr - str1copy;
+	offset = (int)(ptr - str1copy);
 
 	return (char *)(str1 + offset); // Return the original version + offset
 }
@@ -1482,7 +1482,7 @@ int machineInitialize(Machine* machine, UInt8** mainRam, UInt32* mainRamSize, UI
                 int j;
 
                 strcpy(eepromName, machine->slotInfo[i].name);
-                for (j = strlen(eepromName); j > 0 && eepromName[j] != '.'; j--);
+                for (j = (int)strlen(eepromName); j > 0 && eepromName[j] != '.'; j--);
                 eepromName[j] = 0;
                 strcat(eepromName, "_eeprom.rom");
                     
@@ -1542,7 +1542,7 @@ int machineInitialize(Machine* machine, UInt8** mainRam, UInt32* mainRamSize, UI
                 int j;
 
                 strcpy(voiceName, machine->slotInfo[i].name);
-                for (j = strlen(voiceName); j > 0 && voiceName[j] != '.'; j--);
+                for (j = (int)strlen(voiceName); j > 0 && voiceName[j] != '.'; j--);
                 voiceName[j] = 0;
                 strcat(voiceName, "_voice.rom");
                     
@@ -1791,7 +1791,7 @@ int machineInitialize(Machine* machine, UInt8** mainRam, UInt32* mainRamSize, UI
                 int j;
 
                 strcpy(charName, machine->slotInfo[i].name);
-                for (j = strlen(charName); j > 0 && charName[j] != '.'; j--);
+                for (j = (int)strlen(charName); j > 0 && charName[j] != '.'; j--);
                 charName[j] = 0;
                 strcat(charName, "_char.rom");
